activate → pactivate v0.3.5
The very old version of what is now called `pactivate` was breaking on some
systems due to `python` not being available. This latest version handles
this and correctly uses `python3` on those systems.

Also note we add a `./` to the front of the path to pactivate to ensure
we're running the one in this directory, and not one from elsewhere in the
path.